Use pointers to Compatibool everywhere
Because of omitempty, and false being bool's empty value, it's not
possible to send false values unless a pointer is being used.

Also introduce a `fastly.CBool(bool) *Compatibool` helper function that
can be used during the construction of the structs. This helps avoid
having to define intermediate values to get the pointer from.