Add message_type support for s3logging
We'd like to be able to set message_type parameter on s3logging
property.

Potential use case for this is eg. being able to set message_type to
`blank` and use `format` to write logs to s3 as a JSON payload (we can
use `format` to build JSON structured logs).

Most of the tests already included message_type in response body, but it
wasn't being parsed.