fix(request): URL encode names properly
The name of a Fastly resource is used as reference in the URL to GET, PUT, or DELETE a specific
element. The Fastly API accepts all kind of special characters including forward slashes which need
to be URL encoded. Otherwise, the element can't be found in the DB and the Fastly API returns a 404
Not Found error.

fix #65