Add NewRealtimeStatsClientForEndpoint API
This is an RTSClient counterpart to NewClientForEndpoint.

The existing API (NewRealtimeStatsClient) requires FASTLY_API_KEY
to be set in the environment, and panics if the endpoint url
cannot be parsed.

This new API doesn't make any environment requirements and returns
an error if its endpoint cannot be parsed, making it more flexible
for embedding in applications.

Fixes #118