Rename decodeJSON to decodeBodyMap; add decodeMap
This allows JSON-only APIs for round tripping raw JSON output
in the Fastly CLI.

decodeBodyMap has "Body" to indicate that it operates on an
HTTP body, to help signal that it closes its body argument.
And Map indicates that it wants a mapstructure struct.

decodeMap applies our mapstructure defaults to object decoding.

The real work is in client.go; everything else is a rename.