Make "force" flag optional so user has to explicitly opt in
This operation can be "dangerous" if used incorrectly on a subscription
with active domains. In particular, production traffic can disabled.
Despite this operation having no effect if used without the "force"
flag (does not work if subscription is not "issued" and also does not
work if it is "issued", i.e. domains are active), by requiring the flag
the user takes responsibility for this.