Updates needed to support Terraform TLS resources (#259)
* Change TLS certificate `name` attribute to be optional

Custom TLS certificate without names default to the certificates common name or first SAN entry

* Not use uint pointers in ListPrivateKeysInput struct

This was causing problems when trying to use the page[number] parameter so I brought it more in line with the other methods.

* Change TLS Activation `configuration` attribute to be optional (#28)

Activations without the configuration specified default to the default configuration

* Add TLS domains endpoint support

* Drop use of reflection to build request parameters

* Add documentation

* TLS Subscriptions endpoints (#30)

* Add ListTLSSubscriptions function

* Add CreateTLSSubscriptions function

* Add GetTLSSubscriptions function

* Add DeleteTLSSubscriptions function

* Add integration test with all endpoints

* Add TLS Authorizations to TLS Subscriptions (#31)

I had a lot of trouble deserialising the API response into appropriate structs. Initially I could only get it to work by deserialising into a map[string]interface{} but this didn't give the rich typing I wanted to return to the user. After a long time digging through the JSONAPI implementation and an associated PR to allow nested structs (https://github.com/google/jsonapi/pull/99), I found that the go-fastly dependency was on an older version of the library before this PR came in. Updating the version, and playing around some more with the slice of structs instead of slice of pointers, finally got this to work.

Furthermore I found that the Header in ListTLSSubscriptions needed to include the "Accept" header for "application/vnd.api+json" so I added it back in. I previously removed it as it didn't seem to do anything with the Filter and Pagination parameters, but it seems to be required for Include to work.

* Avoid use pointers and uint in ListBulkCertificatesInput struct

* Add allow_untrusted_root field to platform TLS

* Get Certificate ID for TLS Subscriptions when available

The certificate can't really be queried in the same way as the custom TLS certificate, but the ID will be present in the TLS activation automatically created for the TLS subscription, so can help filter for this activation.

* Remove (TLS)\w* from field names of TLS structs

* Add CommonName to Create, Get, and List TLSSubscription functions

* Add "force" flag to TLS Subscription deletion

Allows deleting a subscription with active domains; a potentially
dangerous operation if the subscription handles production traffic. By
using a flag, the user opts in to this risky behaviour and takes
responsibility for knowing what they are doing.

* go mod tidy

* Add comments describing new structs and function

* Document TLSAuthorizations struct

Co-authored-by: Will May <will.j.may@gmail.com>