Add S3 Logging ACL parameter and add Redundancy options
The S3 logging endpoint in the Fastly API supports an `acl` parameter which allows setting the permissions of the uploaded log objects within the bucket.
This commit adds support for this to the SDK.
The Fastly UI console suggests that there are 7 options for this based on the AWS "canned ACL"s, including "private", "public-read" etc.
I have therefore used constants to represent these options, similarly to the redundancy parameters.

Furthermore, there are 4 options for redundancy, but only two were previously supported by the SDK.
I have added the two others for the infrequently accessed storage tiers.