fastly: fix invalid package uploading test
It seems like the expected response from Fastly API slightly changed
since it was added the test for invalid packages. Now, instead of
returning package info with zeroed properties, it properly returns a
specific HTTP error (bad request).

This PR changes the test to be compliant with the new behaviour.

Signed-off-by: Gorka Lerchundi Osa <glertxundi@gmail.com>