add support for Secret Store client keys (#392)
* add support for Secret Store client keys

This lets clients create a client key for local encryption of secrets
before uploading to the Fastly API.

This bumps the minimum required Go version to 1.18 and introduces
`golang.org/x/crypto` as a dependency.

* ci: `go mod download` changed in Go 1.17, don't error on downloads

From the `go help mod download` docs:

> With no arguments, download applies to the modules needed to build and
test the packages in the main module: the modules explicitly required by
the main module if it is at 'go 1.17' or higher, or all
transitively-required modules if at 'go 1.16' or lower.

This causes `go mod tidy` to download additional modules not fetched by
`go mod download`.  The old check errored if there was any output from
`go mod tidy -v`.  So now we omit "downloading" messages.  As a result,
it's also no longer necessary to separately run `go mod download`, as
any modules that are needed are downloaded by `go mod tidy`.

* upgrade staticcheck to support generics (Go 1.18)

Version 0.3.3 (2022.1.3)

* add Encrypt method to ClientKey