Explicitly handle error 404 the API can return.
I'd forgotten about this case, but it's used when the user hasn't set
any Image Optimizer settings, and it's useful for terraform to be able
to easily detect this case.

I assume other users of this API will also appreciate this explicit
error handling, as opposed to having to check for error 404 themselves.