feat(package): add `ClonedFrom` field to PackageMetadata (#527)
* Add ClonedFrom to PackageMetadata

We recently exposed the source repo for e.g. Compute Starter Kits via
the `cloned_from` key in the PackageMetadata resource in the API. This
makes it programmatically accessible.

* Incorporate feedback from PR

* Sort struct fields
* Run `make fix-fixtures FASTLY_TEST_SERVICE_ID=sSsCewrlNQiMzZHiqD8Xr0`

* Update tests

* Update fixtures

This updates the valid.tar.gz package to the current Rust empty Starter
Kit and updates fixtures to match by following the steps in TESTING.md.