breaking(product_enablement): Remove support for NGWAF product. (#560)
Enablement of NGWAF was added recently but never worked, because
enablement of the product requires configuration information to be
included and that was overlooked. Product configuration will be a
significant addition.

While this is technically a breaking change, since it never worked,
and was only used by the Fastly CLI and the Fastly Terraform provider
(both of which will also have NGWAF product enablement removed), there
is no need for a major version number increase.

Support for this product will be re-added in a future version where
'product configuration' is also added.