fix(debug_mode): Fix FASTLY_DEBUG_MODE when used in combination with go-vcr.
While running tests, if go-vcr was enabled (VCR_DISABLE was not set to
'1'), and FASTLY_DEBUG_MODE was set to 'true', and no go-vcr cassettes
existed for the tests being run, then any tests which sent HTTP
request with bodies (PUT, PATCH, etc.) would fail as the combination
of these features would result in an empty body being sent as part of
the request.

This was discovered while running tests for new endpoints, so no
cassettes existed, and FASTLY_DEBUG_MODE was enabled in order to
monitor the actual HTTP requests and responses.