POC for generating API endpoints using code generation
Note: This PR is messy as it includes some refactoring of content from
the 'fastly' package into new 'errors' and 'decodebody' packages;
those will be split out into their own PRs before this is ready for
review.

This PR is a proof-of-concept for generating API endpoints using the
Jennifer code-generation module. It generates endpoints for Brotli
compression and NGWAF as examples. For NGWAF, it includes
configuration-during-enablement for the 'workspace_id'.