Add AllowUntrustedRoot to CreateCustomTLSCertificateInput spec
We have a situation where we would like to upload self-signed certs in a local development flow. The Fastly API supports this attribute, yet it is not present on the official sdk spec.

Although this field is potentially dangerous, it should still be exposed as an option.

In the current situation, we are forced to make a decision between not having a fully localized development flow and hand writing our fastly api interactions. If this change is accepted we can continue to use this fastly sdk as well as have a fully localized development flow.