[CDTOOL-1037] add context param to most inputs, allowing users of the library to pass context to a request (#670)
There are still a few endpoints that you cannot pass context to,
namely any requests that did not already have an input object. Reason
being that adding arguments to those endpoints would be a breaking
change. If there is demand for any of those endpoints to be able to have
context, please open an issue.