fix(domain discovery): add omitempty to response fields and add error handling test (#681)
All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/go-fastly/pulls) for the same
update/change?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why
you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully run tests with your changes locally?

Update the responses to include `omitempty` for json responses. This is
particularly useful when consuming the data from the CLI. The fields
that are not set, should not be displayed, this is how the upstream api
handles the missing fields.

Also added a test for error handling to ensure that the error response
is handled accordingly.

### User Impact

* [x] What is the user impact of this change?

Users should expect the field to not appear instead of `null`

### Are there any considerations that need to be addressed for release?

When marshaling the struct, fields that do not appear will not appear in
the json

---------

Co-authored-by: Anthony Gomez <anthony.gomez@fastly.com>