Allow untrusted root option on CreateCustomTLSCertificateInput (#596)
# Details

The fastly api for creating / uploading custom TLS certificates allows
for passing in the `allow_untrusted_root` flag, indicating that a self
signed CA (untrusted) is permitted to sign the server auth certificates
that get uploaded.

The go-fastly client does not support this flag, presumably because
exposing this option could potentially lead to negative consequences
(someone uploading a self signed cert that browsers don't trust for a
domain that is already seeing production traffic).

Without this change, we (SeatGeek) am unable to have a fully localized
test environment leveraging a self signed CA that I create on the fly to
test / validate our integration with the fastly custom tls cert upload
process.

## Rationale

* This attribute `allow_untrusted_root` is available on the API, so it
should also be present on the official sdk.
* This attribute is off by default, with a full description of the
drawbacks to using this flag.
* Even if clients were to set this flag, **this feature is still gated
behind an account level flag** that must be turned on by someone
internal to Fastly.

Given the multiple layers of protection and the "default off" behavior
of this flag, I think that this attribute can be safely reflected in the
schema without consequence to customers

Without this, we wont be able to leverage the go-fastly client which
would be a bummer.

All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/go-fastly/pulls) for the same
update/change?

<!-- You can erase any parts of this template not applicable to your
Pull Request. -->

### New Feature Submissions:

* [x] Does your submission pass tests?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why
you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully run tests with your changes locally?

### User Impact

* [x] What is the user impact of this change?

This change is backwards compatible with existing behavior.

This change will allow clients to enable `allow_untrusted_root` when
uploading a custom TLS cert

### Are there any considerations that need to be addressed for release?

<!-- Any breaking changes, etc -->
None

---------

Co-authored-by: Kevin P. Fleming <kpfleming@users.noreply.github.com>