feat(client): Add support for impersonation.
A new 'impersonation' package provides a function which can be used to
construct a context.Context object containing a Fastly CID. If that
context is passed to any of go-fastly's API functions which generate
HTTP requests to the Fastly API endpoint, the contained CID will be
included in the request.

As noted in the package documentation, impersonation is only available
to Fastly employees, it is not usable by customers.