feat(client): Add support for impersonation. (#724)
A new 'impersonation' package provides a function which can be used to
construct a context.Context object containing a Fastly CID. If that
context is passed to any of go-fastly's API functions which generate
HTTP requests to the Fastly API endpoint, the contained CID will be
included in the request.

As noted in the package documentation, impersonation is only available
to Fastly employees, it is not usable by customers.

 All Submissions:

* [X] Have you followed the guidelines in our Contributing document?
* [X] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/go-fastly/pulls) for the same
update/change?

<!-- You can erase any parts of this template not applicable to your
Pull Request. -->

### New Feature Submissions:

* [X] Does your submission pass tests?

### Changes to Core Features:

* [X] Have you added an explanation of what your changes do and why
you'd like us to include them?
* [X] Have you written new tests for your core changes, as applicable?
* [X] Have you successfully run tests with your changes locally?