[CDTOOL-1111] Support for compression codec and gzip for https endpoint (#734)
All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/go-fastly/pulls) for the same
update/change?

<!-- You can erase any parts of this template not applicable to your
Pull Request. -->

### New Feature Submissions:

* [x ] Does your submission pass tests?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why
you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully run tests with your changes locally?

### User Impact

* [x] What is the user impact of this change? - Support for compression
has been added to the https endpoint and exists in the UI and API.
Adding it to go-fastly will allow us to add it to cli and terraform

### Are there any considerations that need to be addressed for release?
This change gates adding support to the CLI and terraform