Remove package-level godoc about ERL being an optional platform feature.
This doesn't belong here, as this file is part of the 'fastly' package
and this comment does not apply to the entire package. In addition
there are many other API functions in this module which only work if
optional platform features are purchased by the customer and they
don't have cautionary comments like this.