Remove package-level godoc about ERL being an optional platform feature. (#744)
This doesn't belong here, as this file is part of the 'fastly' package
and this comment does not apply to the entire package. In addition there
are many other API functions in this module which only work if optional
platform features are purchased by the customer and they don't have
cautionary comments like this.

 All Submissions:

* [X] Have you followed the guidelines in our Contributing document?
* [X] Have you checked to ensure there aren't other open [Pull
Requests](https://github.com/fastly/go-fastly/pulls) for the same
update/change?